+++ ceammc/ext/tests/random/test_random_float.cpp
@@ -30,6 +30,7 @@ TEST_CASE("random.float", "[externals]")
 
             REQUIRE_PROPERTY(t, @min, 0);
             REQUIRE_PROPERTY(t, @max, 1);
+            REQUIRE_PROPERTY(t, @seed, 0);
         }
 
         SECTION("args max")
@@ -80,5 +81,33 @@ TEST_CASE("random.float", "[externals]")
             REQUIRE(f >= 3);
             REQUIRE(f < 4);
         }
+    }
+
+    SECTION("seed")
+    {
+        TExt t("random.f", "@seed", 1);
+        constexpr size_t N = 10;
+
+        std::vector<t_float> v0;
+        for (int i = 0; i < N; i++) {
+            t.bang();
+            v0.push_back(floatAt(t));
+        }
+
+        std::vector<t_float> v1;
+        for (int i = 0; i < N; i++) {
+            t.bang();
+            v1.push_back(floatAt(t));
+        }
+
+        REQUIRE(v0 != v1);
+
+        t->setProperty("@seed", LF(1));
+        for (int i = 0; i < N; i++) {
+            t.bang();
+            v1[i] = floatAt(t);
+        }
+
+        REQUIRE(v0 == v1);
     }
 }