+++ ceammc/CHANGELOG.md
@@ -6,6 +6,7 @@
 - patch.deps added - list current patch dependencies
 - ceammc.search added
 - random.int @seed property added to gen reproducible random sequences
+- random.float @seed property added to gen reproducible random sequences
 
 ### Changed:
 - ext_info output format changed.